🚀 Major Entity Notes Integration Improvements
This comprehensive update brings significant enhancements to the Entity Notes integration:

🔧 **Core Improvements:**
- **Automatic upgrade handling**: Seamlessly upgrades existing installations while preserving user data
- **Auto-reload on config changes**: Integration automatically reloads when settings are modified
- **Cache-busting system**: Ensures JavaScript updates are immediately reflected in browsers
- **Enhanced error handling**: More robust error handling and logging throughout

📱 **User Interface Enhancements:**
- **Dynamic configuration injection**: JavaScript receives backend settings automatically
- **Improved visual feedback**: Better character counting and validation
- **Enhanced dialog injection**: More reliable entity dialog detection and notes insertion
- **Responsive design improvements**: Better handling of various screen sizes

⚙️ **Configuration & Manifest Updates:**
- **Proper integration_type**: Fixed manifest.json to properly identify as "helper"
- **Custom icon support**: Added proper icon.png for integration branding
- **Expanded configuration options**: Added debug logging, auto-backup, and max length settings
- **Complete translations**: Added comprehensive English translations

🛠️ **Technical Enhancements:**
- **Service registration**: Full suite of services for note management (set, get, delete, list, backup, restore)
- **Event system**: Proper event firing for note updates
- **Storage optimization**: Improved data persistence and backup capabilities
- **Debug capabilities**: Comprehensive logging and debugging features

This update maintains full backward compatibility while significantly improving functionality, reliability, and user experience.
