+++ custom_components/entity_notes/__init__.py
@@ -1,6 +1,8 @@
+
 """Entity Notes integration for Home Assistant."""
 import logging
 import voluptuous as vol
+import time
 from homeassistant.core import HomeAssistant
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers.typing import ConfigType
@@ -10,109 +12,261 @@ from homeassistant.components.frontend import add_extra_js_url
 from aiohttp import web
 import json
 import os
+from pathlib import Path
+
+from .const import (
+    DOMAIN,
+    STORAGE_VERSION,
+    STORAGE_KEY,
+    MAX_NOTE_LENGTH,
+    CONF_DEBUG_LOGGING,
+    CONF_MAX_NOTE_LENGTH,
+    CONF_AUTO_BACKUP,
+    DEFAULT_DEBUG_LOGGING,
+    DEFAULT_MAX_NOTE_LENGTH,
+    DEFAULT_AUTO_BACKUP,
+    FRONTEND_JS_PATH,
+    EVENT_NOTES_UPDATED,
+    SERVICE_SET_NOTE,
+    SERVICE_GET_NOTE,
+    SERVICE_DELETE_NOTE,
+    SERVICE_LIST_NOTES,
+    SERVICE_BACKUP_NOTES,
+    SERVICE_RESTORE_NOTES,
+)
 
 _LOGGER = logging.getLogger(__name__)
 
-DOMAIN = "entity_notes"
-STORAGE_VERSION = 1
-STORAGE_KEY = "entity_notes.notes"
-
-# Module-level debug log to see if this file is being loaded
-_LOGGER.warning("🔧 DEBUG: __init__.py module loaded for domain: %s", DOMAIN)
-
 CONFIG_SCHEMA = vol.Schema({DOMAIN: vol.Schema({})}, extra=vol.ALLOW_EXTRA)
+
 
 async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
     """Set up the Entity Notes integration from configuration.yaml."""
-    _LOGGER.warning("🔧 DEBUG: async_setup called with config: %s", config)
     # This is for backward compatibility with configuration.yaml
     # The main setup now happens in async_setup_entry
-    _LOGGER.warning("🔧 DEBUG: async_setup completed successfully")
     return True
+
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up Entity Notes from a config entry."""
-    _LOGGER.warning("🔧 DEBUG: async_setup_entry called with entry: %s", entry)
-    _LOGGER.warning("🔧 DEBUG: Entry data: %s", entry.data)
-    _LOGGER.warning("🔧 DEBUG: Entry domain: %s", entry.domain)
-    _LOGGER.warning("🔧 DEBUG: Entry title: %s", entry.title)
+    _LOGGER.info("Setting up Entity Notes integration")
+    
+    # Get configuration options
+    options = entry.options or {}
+    debug_logging = options.get(CONF_DEBUG_LOGGING, DEFAULT_DEBUG_LOGGING)
+    max_note_length = options.get(CONF_MAX_NOTE_LENGTH, DEFAULT_MAX_NOTE_LENGTH)
+    auto_backup = options.get(CONF_AUTO_BACKUP, DEFAULT_AUTO_BACKUP)
+    
+    if debug_logging:
+        _LOGGER.setLevel(logging.DEBUG)
+        _LOGGER.debug("Debug logging enabled for Entity Notes")
     
     try:
-        _LOGGER.warning("🔧 DEBUG: Starting Entity Notes integration setup")
-
         # Initialize storage
-        _LOGGER.warning("🔧 DEBUG: Initializing storage")
+        _LOGGER.debug("Initializing storage")
         store = Store(hass, STORAGE_VERSION, STORAGE_KEY)
 
         # Load existing notes
         try:
-            _LOGGER.warning("🔧 DEBUG: Loading existing notes from storage")
             stored_data = await store.async_load()
             notes_data = stored_data or {}
-            _LOGGER.warning("🔧 DEBUG: Loaded %d existing notes", len(notes_data))
+            _LOGGER.debug("Loaded %d existing notes", len(notes_data))
         except Exception as e:
-            _LOGGER.warning("🔧 DEBUG: Could not load notes storage, starting fresh: %s", e)
+            _LOGGER.warning("Could not load notes storage, starting fresh: %s", e)
             notes_data = {}
 
-        # Store the storage instance and data in hass.data
-        _LOGGER.warning("🔧 DEBUG: Setting up hass.data for domain %s", DOMAIN)
+        # Generate cache buster for JavaScript file
+        cache_buster = int(time.time())
+
+        # Store the configuration and data in hass.data
         hass.data.setdefault(DOMAIN, {})
         hass.data[DOMAIN] = {
             "store": store,
-            "notes": notes_data
+            "notes": notes_data,
+            "config": {
+                CONF_DEBUG_LOGGING: debug_logging,
+                CONF_MAX_NOTE_LENGTH: max_note_length,
+                CONF_AUTO_BACKUP: auto_backup,
+            },
+            "entry_id": entry.entry_id,
+            "cache_buster": cache_buster,
         }
-        _LOGGER.warning("🔧 DEBUG: hass.data setup complete")
 
         # Register the API view
-        _LOGGER.warning("🔧 DEBUG: Registering EntityNotesView")
         hass.http.register_view(EntityNotesView())
-        _LOGGER.warning("🔧 DEBUG: EntityNotesView registered successfully")
+        _LOGGER.debug("EntityNotesView registered")
         
         # Register the JavaScript file serving view
-        _LOGGER.warning("🔧 DEBUG: Registering EntityNotesJSView")
         hass.http.register_view(EntityNotesJSView())
-        _LOGGER.warning("🔧 DEBUG: EntityNotesJSView registered successfully")
+        _LOGGER.debug("EntityNotesJSView registered")
 
-        # Register frontend resource
-        # Serve the JS file directly from the integration
+        # Register frontend resource with cache busting
         try:
-            _LOGGER.warning("🔧 DEBUG: Registering frontend resource")
-            # Register the JS file served by our own endpoint
-            add_extra_js_url(hass, "/api/entity_notes/entity-notes.js")
-            _LOGGER.warning("🔧 DEBUG: Frontend resource registered: /api/entity_notes/entity-notes.js")
-                
+            js_url_with_cache_buster = f"/api/entity_notes/entity-notes.js?v={cache_buster}"
+            add_extra_js_url(hass, js_url_with_cache_buster)
+            _LOGGER.debug("Frontend resource registered: %s", js_url_with_cache_buster)
         except Exception as e:
-            _LOGGER.error("🔧 DEBUG: Failed to register frontend resource: %s", e)
+            _LOGGER.error("Failed to register frontend resource: %s", e)
 
-        _LOGGER.warning("🔧 DEBUG: Entity Notes integration setup complete successfully")
+        # Register services
+        await async_register_services(hass)
+
+        _LOGGER.info("Entity Notes integration setup completed successfully")
         return True
         
     except Exception as e:
-        _LOGGER.error("🔧 DEBUG: CRITICAL ERROR in async_setup_entry: %s", e)
-        _LOGGER.error("🔧 DEBUG: Exception type: %s", type(e).__name__)
-        import traceback
-        _LOGGER.error("🔧 DEBUG: Full traceback: %s", traceback.format_exc())
+        _LOGGER.error("Critical error in Entity Notes setup: %s", e)
         return False
+
 
 async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Unload a config entry."""
-    # Clean up if needed
-    return True
+    _LOGGER.info("Unloading Entity Notes integration")
+    
+    try:
+        # Remove services
+        services_to_remove = [
+            SERVICE_SET_NOTE,
+            SERVICE_GET_NOTE,
+            SERVICE_DELETE_NOTE,
+            SERVICE_LIST_NOTES,
+            SERVICE_BACKUP_NOTES,
+            SERVICE_RESTORE_NOTES,
+        ]
+        
+        for service in services_to_remove:
+            if hass.services.has_service(DOMAIN, service):
+                hass.services.async_remove(DOMAIN, service)
+        
+        # Clean up data
+        hass.data.pop(DOMAIN, None)
+        
+        _LOGGER.info("Entity Notes integration unloaded successfully")
+        return True
+        
+    except Exception as e:
+        _LOGGER.error("Error unloading Entity Notes: %s", e)
+        return False
+
+
+async def async_register_services(hass: HomeAssistant) -> None:
+    """Register the Entity Notes services."""
+    
+    async def set_note_service(call):
+        """Set a note for an entity."""
+        entity_id = call.data.get("entity_id")
+        note = call.data.get("note", "")
+        
+        if not entity_id:
+            _LOGGER.error("No entity_id provided for set_note service")
+            return
+            
+        store = hass.data[DOMAIN]["store"]
+        notes_data = hass.data[DOMAIN]["notes"]
+        max_length = hass.data[DOMAIN]["config"][CONF_MAX_NOTE_LENGTH]
+        
+        # Truncate note if too long
+        if len(note) > max_length:
+            note = note[:max_length]
+            _LOGGER.warning("Note truncated to %d characters for %s", max_length, entity_id)
+        
+        if note.strip():
+            notes_data[entity_id] = note.strip()
+            _LOGGER.info("Set note for %s", entity_id)
+        else:
+            notes_data.pop(entity_id, None)
+            _LOGGER.info("Removed note for %s", entity_id)
+        
+        await store.async_save(notes_data)
+        hass.bus.async_fire(EVENT_NOTES_UPDATED, {"entity_id": entity_id, "note": note})
+
+    async def get_note_service(call):
+        """Get a note for an entity."""
+        entity_id = call.data.get("entity_id")
+        if not entity_id:
+            return
+            
+        notes_data = hass.data[DOMAIN]["notes"]
+        note = notes_data.get(entity_id, "")
+        
+        hass.bus.async_fire("entity_notes_get_response", {
+            "entity_id": entity_id,
+            "note": note
+        })
+
+    async def delete_note_service(call):
+        """Delete a note for an entity."""
+        entity_id = call.data.get("entity_id")
+        if not entity_id:
+            return
+            
+        store = hass.data[DOMAIN]["store"]
+        notes_data = hass.data[DOMAIN]["notes"]
+        
+        if entity_id in notes_data:
+            del notes_data[entity_id]
+            await store.async_save(notes_data)
+            hass.bus.async_fire(EVENT_NOTES_UPDATED, {"entity_id": entity_id, "note": ""})
+            _LOGGER.info("Deleted note for %s", entity_id)
+
+    async def list_notes_service(call):
+        """List all notes."""
+        notes_data = hass.data[DOMAIN]["notes"]
+        hass.bus.async_fire("entity_notes_list_response", {"notes": dict(notes_data)})
+
+    async def backup_notes_service(call):
+        """Backup all notes to a file."""
+        notes_data = hass.data[DOMAIN]["notes"]
+        backup_path = hass.config.path("entity_notes_backup.json")
+        
+        try:
+            with open(backup_path, "w") as f:
+                json.dump(notes_data, f, indent=2)
+            _LOGGER.info("Notes backed up to %s", backup_path)
+        except Exception as e:
+            _LOGGER.error("Failed to backup notes: %s", e)
+
+    async def restore_notes_service(call):
+        """Restore notes from a backup file."""
+        backup_path = hass.config.path("entity_notes_backup.json")
+        
+        try:
+            with open(backup_path, "r") as f:
+                backup_data = json.load(f)
+            
+            store = hass.data[DOMAIN]["store"]
+            hass.data[DOMAIN]["notes"].update(backup_data)
+            await store.async_save(hass.data[DOMAIN]["notes"])
+            _LOGGER.info("Notes restored from %s", backup_path)
+        except Exception as e:
+            _LOGGER.error("Failed to restore notes: %s", e)
+
+    # Register services
+    hass.services.async_register(DOMAIN, SERVICE_SET_NOTE, set_note_service)
+    hass.services.async_register(DOMAIN, SERVICE_GET_NOTE, get_note_service)
+    hass.services.async_register(DOMAIN, SERVICE_DELETE_NOTE, delete_note_service)
+    hass.services.async_register(DOMAIN, SERVICE_LIST_NOTES, list_notes_service)
+    hass.services.async_register(DOMAIN, SERVICE_BACKUP_NOTES, backup_notes_service)
+    hass.services.async_register(DOMAIN, SERVICE_RESTORE_NOTES, restore_notes_service)
+
 
 class EntityNotesView(HomeAssistantView):
     """Handle Entity Notes API requests."""
 
     url = "/api/entity_notes/{entity_id}"
     name = "api:entity_notes"
-    requires_auth = False  # Simplified - no auth required for local requests
+    requires_auth = False  # Local requests only
 
     async def get(self, request, entity_id):
         """Get note for an entity."""
         hass = request.app["hass"]
         notes_data = hass.data[DOMAIN]["notes"]
-
         note = notes_data.get(entity_id, "")
-        _LOGGER.debug("Retrieved note for %s: %s", entity_id, note[:50] + "..." if len(note) > 50 else note)
+        
+        debug_logging = hass.data[DOMAIN]["config"][CONF_DEBUG_LOGGING]
+        if debug_logging:
+            _LOGGER.debug("Retrieved note for %s: %s", entity_id, note[:50] + "..." if len(note) > 50 else note)
+            
         return web.json_response({"note": note})
 
     async def post(self, request, entity_id):
@@ -120,21 +274,32 @@ class EntityNotesView(HomeAssistantView):
         hass = request.app["hass"]
         store = hass.data[DOMAIN]["store"]
         notes_data = hass.data[DOMAIN]["notes"]
+        max_length = hass.data[DOMAIN]["config"][CONF_MAX_NOTE_LENGTH]
+        debug_logging = hass.data[DOMAIN]["config"][CONF_DEBUG_LOGGING]
 
         try:
             data = await request.json()
             note = data.get("note", "").strip()
 
+            # Enforce max length
+            if len(note) > max_length:
+                note = note[:max_length]
+
             if note:
                 notes_data[entity_id] = note
-                _LOGGER.info("Saved note for %s: %s", entity_id, note[:50] + "..." if len(note) > 50 else note)
+                if debug_logging:
+                    _LOGGER.debug("Saved note for %s: %s", entity_id, note[:50] + "..." if len(note) > 50 else note)
             else:
                 # Remove empty notes
                 notes_data.pop(entity_id, None)
-                _LOGGER.info("Removed empty note for %s", entity_id)
+                if debug_logging:
+                    _LOGGER.debug("Removed empty note for %s", entity_id)
 
             # Save to persistent storage
             await store.async_save(notes_data)
+            
+            # Fire event
+            hass.bus.async_fire(EVENT_NOTES_UPDATED, {"entity_id": entity_id, "note": note})
 
             return web.json_response({"status": "success"})
 
@@ -147,14 +312,19 @@ class EntityNotesView(HomeAssistantView):
         hass = request.app["hass"]
         store = hass.data[DOMAIN]["store"]
         notes_data = hass.data[DOMAIN]["notes"]
+        debug_logging = hass.data[DOMAIN]["config"][CONF_DEBUG_LOGGING]
 
         try:
             if entity_id in notes_data:
                 del notes_data[entity_id]
-                _LOGGER.info("Deleted note for %s", entity_id)
+                if debug_logging:
+                    _LOGGER.debug("Deleted note for %s", entity_id)
 
                 # Save to persistent storage
                 await store.async_save(notes_data)
+                
+                # Fire event
+                hass.bus.async_fire(EVENT_NOTES_UPDATED, {"entity_id": entity_id, "note": ""})
 
                 return web.json_response({"status": "deleted"})
             else:
@@ -164,6 +334,7 @@ class EntityNotesView(HomeAssistantView):
             _LOGGER.error("Error deleting note for %s: %s", entity_id, e)
             return web.json_response({"error": str(e)}, status=500)
 
+
 class EntityNotesJSView(HomeAssistantView):
     """Serve the Entity Notes JavaScript file."""
 
@@ -171,486 +342,28 @@ class EntityNotesJSView(HomeAssistantView):
     name = "api:entity_notes_js"
     requires_auth = False
 
-    # Embedded JavaScript content to avoid file I/O operations
-    JS_CONTENT = """console.log('Entity Notes: Script loading...');
-
-// Create global namespace with configurable debug mode
-window.entityNotes = {
-    version: '1.2.0-ui-improved',
-    debug: false,  // Default: quiet mode for production
-    
-    // Convenience methods for users
-    enableDebug: function() { 
-        this.debug = true; 
-        console.log('Entity Notes: Debug mode enabled. Refresh page or open entity dialogs to see debug output.');
-    },
-    disableDebug: function() { 
-        this.debug = false; 
-        console.log('Entity Notes: Debug mode disabled.');
-    }
-};
-
-// Debug logging function - only logs when debug mode is enabled
-function debugLog(message) {
-    if (window.entityNotes.debug) {
-        console.log(message);
-    }
-}
-
-// Always log critical messages (errors, warnings, success)
-function infoLog(message) {
-    console.log(message);
-}
-
-class EntityNotesCard extends HTMLElement {
-    constructor() {
-        super();
-        this.attachShadow({ mode: 'open' });
-        debugLog('Entity Notes: EntityNotesCard constructor called');
-    }
-
-    connectedCallback() {
-        debugLog('Entity Notes: EntityNotesCard connected');
-        this.render();
-        this.setupEventListeners();
-    }
-
-    render() {
-        this.shadowRoot.innerHTML = `
-            <style>
-                .entity-notes-container {
-                    margin: 8px 0;
-                    padding: 8px;
-                    border: none;
-                    border-radius: 4px;
-                    background: transparent;
-                }
-                .entity-notes-textarea {
-                    width: 100%;
-                    min-height: 36px;
-                    max-height: 300px;
-                    padding: 6px 8px;
-                    border: 1px solid var(--divider-color, #e0e0e0);
-                    border-radius: 4px;
-                    background: var(--primary-background-color, white);
-                    color: var(--primary-text-color, black);
-                    font-family: inherit;
-                    font-size: 14px;
-                    line-height: 1.4;
-                    resize: none;
-                    overflow: hidden;
-                    box-sizing: border-box;
-                    transition: height 0.1s ease;
-                    outline: none;
-                }
-                .entity-notes-textarea:focus {
-                    border-color: var(--primary-color, #03a9f4);
-                    box-shadow: 0 0 0 1px var(--primary-color, #03a9f4);
-                }
-                .entity-notes-actions {
-                    display: flex;
-                    gap: 8px;
-                    margin-top: 8px;
-                    justify-content: flex-end;
-                }
-                .entity-notes-button {
-                    padding: 6px 12px;
-                    border: none;
-                    border-radius: 4px;
-                    cursor: pointer;
-                    font-size: 12px;
-                    font-weight: 500;
-                    text-transform: uppercase;
-                }
-                .entity-notes-save {
-                    background: var(--primary-color, #03a9f4);
-                    color: white;
-                }
-                .entity-notes-delete {
-                    background: var(--error-color, #f44336);
-                    color: white;
-                }
-                .entity-notes-char-count {
-                    font-size: 11px;
-                    color: var(--secondary-text-color, #666);
-                    margin-top: 4px;
-                    text-align: right;
-                }
-                .entity-notes-char-count.warning {
-                    color: var(--warning-color, #ff9800);
-                }
-                .entity-notes-char-count.error {
-                    color: var(--error-color, #f44336);
-                }
-            </style>
-            <div class="entity-notes-container">
-                <textarea 
-                    class="entity-notes-textarea" 
-                    placeholder="Notes"
-                    maxlength="200"
-                    rows="1"
-                ></textarea>
-                <div class="entity-notes-char-count">0/200</div>
-                <div class="entity-notes-actions">
-                    <button class="entity-notes-button entity-notes-delete">DELETE</button>
-                    <button class="entity-notes-button entity-notes-save">SAVE</button>
-                </div>
-            </div>
-        `;
-    }
-
-    setupEventListeners() {
-        const textarea = this.shadowRoot.querySelector('.entity-notes-textarea');
-        const charCount = this.shadowRoot.querySelector('.entity-notes-char-count');
-        const saveBtn = this.shadowRoot.querySelector('.entity-notes-save');
-        const deleteBtn = this.shadowRoot.querySelector('.entity-notes-delete');
-
-        textarea.addEventListener('input', () => {
-            this.updateCharCount();
-            this.autoResize();
-        });
-
-        textarea.addEventListener('focus', () => {
-            this.autoResize();
-        });
-
-        saveBtn.addEventListener('click', () => this.saveNote());
-        deleteBtn.addEventListener('click', () => this.deleteNote());
-    }
-
-    updateCharCount() {
-        const textarea = this.shadowRoot.querySelector('.entity-notes-textarea');
-        const charCount = this.shadowRoot.querySelector('.entity-notes-char-count');
-        const count = textarea.value.length;
-        
-        charCount.textContent = `${count}/200`;
-        
-        charCount.classList.remove('warning', 'error');
-        if (count > 180) charCount.classList.add('warning');
-        if (count > 200) charCount.classList.add('error');
-    }
-
-    autoResize() {
-        const textarea = this.shadowRoot.querySelector('.entity-notes-textarea');
-        textarea.style.height = 'auto';
-        const newHeight = Math.max(40, Math.min(300, textarea.scrollHeight));
-        textarea.style.height = newHeight + 'px';
-    }
-
-    async loadNote() {
-        const entityId = this.getAttribute('entity-id');
-        debugLog('Entity Notes: Loading note for ' + entityId);
-        if (!entityId) return;
-
-        try {
-            const response = await fetch(`/api/entity_notes/${entityId}`);
-            const data = await response.json();
-            
-            const textarea = this.shadowRoot.querySelector('.entity-notes-textarea');
-            textarea.value = data.note || '';
-            
-            this.updateCharCount();
-            setTimeout(() => this.autoResize(), 10);
-            
-        } catch (error) {
-            console.error('Entity Notes: Error loading note:', error);
-        }
-    }
-
-    async saveNote() {
-        const entityId = this.getAttribute('entity-id');
-        const textarea = this.shadowRoot.querySelector('.entity-notes-textarea');
-        const note = textarea.value.trim();
-
-        debugLog('Entity Notes: Saving note for ' + entityId + ': ' + note);
-
-        try {
-            const response = await fetch(`/api/entity_notes/${entityId}`, {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ note })
-            });
-
-            if (response.ok) {
-                debugLog('Entity Notes: Note saved successfully');
-            } else {
-                console.error('Entity Notes: Save failed');
-            }
-        } catch (error) {
-            console.error('Entity Notes: Error saving note:', error);
-        }
-    }
-
-    async deleteNote() {
-        const entityId = this.getAttribute('entity-id');
-        debugLog('Entity Notes: Deleting note for ' + entityId);
-
-        try {
-            const response = await fetch(`/api/entity_notes/${entityId}`, {
-                method: 'DELETE'
-            });
-
-            if (response.ok) {
-                const textarea = this.shadowRoot.querySelector('.entity-notes-textarea');
-                textarea.value = '';
-                this.updateCharCount();
-                this.autoResize();
-                debugLog('Entity Notes: Note deleted successfully');
-            }
-        } catch (error) {
-            console.error('Entity Notes: Error deleting note:', error);
-        }
-    }
-}
-
-// Register both element names for compatibility
-if (!customElements.get('entity-notes-card')) {
-    customElements.define('entity-notes-card', EntityNotesCard);
-    infoLog('Entity Notes: Integration loaded successfully');
-}
-
-// Wrap second registration in try-catch to prevent script execution from stopping
-if (!customElements.get('entity-notes')) {
-    try {
-        customElements.define('entity-notes', EntityNotesCard);
-        debugLog('Entity Notes: Secondary element registration successful');
-    } catch (error) {
-        debugLog('Entity Notes: Secondary element registration failed (this is OK): ' + error.message);
-        // This is expected - browsers don't allow the same constructor for multiple element names
-        // The first registration (entity-notes-card) is sufficient for functionality
-    }
-}
-
-// Store reference for debugging
-window.entityNotes.EntityNotesCard = EntityNotesCard;
-
-function findEntityId(dialog) {
-    debugLog('Entity Notes: Finding entity ID for dialog');
-    
-    // Try multiple methods to get entity ID
-    const methods = [
-        () => dialog.stateObj?.entity_id,
-        () => dialog._stateObj?.entity_id,
-        () => dialog.entityId,
-        () => dialog._entityId,
-        () => dialog.getAttribute?.('entity-id'),
-        () => dialog.dataset?.entityId,
-        () => {
-            const stateObj = dialog.querySelector?.('[state-obj]')?.stateObj;
-            return stateObj?.entity_id;
-        }
-    ];
-    
-    for (const method of methods) {
-        try {
-            const entityId = method();
-            if (entityId) {
-                debugLog('Entity Notes: Found entity ID: ' + entityId);
-                return entityId;
-            }
-        } catch (e) {
-            // Continue to next method
-        }
-    }
-    
-    debugLog('Entity Notes: No entity ID found');
-    return null;
-}
-
-function injectNotesIntoDialog(dialog) {
-    debugLog('Entity Notes: Attempting to inject notes into dialog');
-    
-    if (!dialog || !dialog.shadowRoot) {
-        debugLog('Entity Notes: No dialog or shadowRoot found');
-        return;
-    }
-    
-    // Check if already injected
-    if (dialog.shadowRoot.querySelector('entity-notes-card')) {
-        debugLog('Entity Notes: Notes already injected');
-        return;
-    }
-    
-    const entityId = findEntityId(dialog);
-    if (!entityId) {
-        debugLog('Entity Notes: No entity ID found for dialog');
-        return;
-    }
-    
-    // Try multiple selectors to find content area
-    const selectors = [
-        '.content',
-        '.mdc-dialog__content',
-        '[slot="content"]',
-        '.dialog-content',
-        'ha-dialog-content'
-    ];
-    
-    let contentArea = null;
-    for (const selector of selectors) {
-        contentArea = dialog.shadowRoot.querySelector(selector);
-        if (contentArea) {
-            debugLog('Entity Notes: Found content area with selector: ' + selector);
-            break;
-        }
-    }
-    
-    if (!contentArea) {
-        debugLog('Entity Notes: No content area found');
-        return;
-    }
-    
-    // Create and inject notes card
-    const notesCard = document.createElement('entity-notes-card');
-    notesCard.setAttribute('entity-id', entityId);
-    contentArea.appendChild(notesCard);
-    
-    debugLog('Entity Notes: Notes card injected for entity: ' + entityId);
-    
-    // Load the note after a short delay
-    setTimeout(() => {
-        notesCard.loadNote();
-    }, 100);
-}
-
-function setupDialogObserver() {
-    debugLog('Entity Notes: Setting up dialog observer');
-    
-    const homeAssistant = document.querySelector('home-assistant');
-    if (!homeAssistant?.shadowRoot) {
-        debugLog('Entity Notes: Home Assistant shadow root not found, retrying in 1 second...');
-        setTimeout(setupDialogObserver, 1000);
-        return;
-    }
-    
-    // Observer specifically for Home Assistant shadow DOM
-    const observer = new MutationObserver((mutations) => {
-        mutations.forEach((mutation) => {
-            mutation.addedNodes.forEach((node) => {
-                if (node.nodeType === 1) {
-                    debugLog('Entity Notes: Node added to shadow DOM: ' + node.tagName);
-                    
-                    // Check if this is a more-info dialog
-                    if (node.tagName === 'HA-MORE-INFO-DIALOG') {
-                        debugLog('Entity Notes: More-info dialog detected');
-                        
-                        // Try injection with multiple delays to ensure dialog is fully loaded
-                        [100, 300, 600, 1000].forEach(delay => {
-                            setTimeout(() => {
-                                debugLog('Entity Notes: Attempting injection after ' + delay + 'ms delay');
-                                injectNotesIntoDialog(node);
-                            }, delay);
-                        });
-                    }
-                    
-                    // Also check for nested dialogs
-                    const nestedDialogs = node.querySelectorAll?.('ha-more-info-dialog');
-                    nestedDialogs?.forEach(dialog => {
-                        debugLog('Entity Notes: Found nested dialog');
-                        [100, 300, 600].forEach(delay => {
-                            setTimeout(() => injectNotesIntoDialog(dialog), delay);
-                        });
-                    });
-                }
-            });
-        });
-    });
-    
-    // Observe the Home Assistant shadow root (where dialogs are actually created)
-    debugLog('Entity Notes: Observing home-assistant shadow root');
-    observer.observe(homeAssistant.shadowRoot, { 
-        childList: true, 
-        subtree: true 
-    });
-    
-    // Also check for existing dialogs in shadow root
-    const existingDialogs = homeAssistant.shadowRoot.querySelectorAll('ha-more-info-dialog');
-    if (existingDialogs.length > 0) {
-        debugLog('Entity Notes: Found existing dialogs in shadow root: ' + existingDialogs.length);
-        existingDialogs.forEach(dialog => {
-            setTimeout(() => injectNotesIntoDialog(dialog), 100);
-        });
-    }
-    
-    window.entityNotes.observer = observer;
-    infoLog('Entity Notes: Observer setup complete');
-}
-
-// Initialize when DOM is ready
-function initialize() {
-    debugLog('Entity Notes: Initializing...');
-    setupDialogObserver();
-    
-    // Try to inject into any existing dialogs in shadow DOM
-    const homeAssistant = document.querySelector('home-assistant');
-    if (homeAssistant?.shadowRoot) {
-        const existingDialogs = homeAssistant.shadowRoot.querySelectorAll('ha-more-info-dialog');
-        debugLog('Entity Notes: Found existing dialogs during init: ' + existingDialogs.length);
-        existingDialogs.forEach(dialog => {
-            setTimeout(() => injectNotesIntoDialog(dialog), 100);
-        });
-    }
-    
-    debugLog('Entity Notes: Initialization complete');
-}
-
-// Make initialize function globally accessible for debugging
-window.entityNotes.initialize = initialize;
-
-// Robust initialization with error handling
-function initializeWithErrorHandling() {
-    try {
-        debugLog('Entity Notes: Starting initialization...');
-        debugLog('Entity Notes: DOM ready state: ' + document.readyState);
-        
-        initialize();
-        
-        infoLog('Entity Notes: Initialization completed successfully');
-    } catch (error) {
-        console.error('Entity Notes: Initialization failed:', error);
-        console.error('Entity Notes: Error stack:', error.stack);
-        
-        // Try fallback initialization after delay
-        setTimeout(() => {
-            try {
-                debugLog('Entity Notes: Attempting fallback initialization...');
-                initialize();
-            } catch (fallbackError) {
-                console.error('Entity Notes: Fallback initialization also failed:', fallbackError);
-            }
-        }, 3000);
-    }
-}
-
-// Handle all DOM ready scenarios
-if (document.readyState === 'loading') {
-    debugLog('Entity Notes: DOM still loading, waiting for DOMContentLoaded...');
-    document.addEventListener('DOMContentLoaded', initializeWithErrorHandling);
-} else {
-    debugLog('Entity Notes: DOM already ready, initializing immediately...');
-    // Use setTimeout to ensure this runs after the current execution stack
-    setTimeout(initializeWithErrorHandling, 0);
-}
-
-// Additional fallback: try initialization after a delay regardless
-setTimeout(() => {
-    if (!window.entityNotes.observer) {
-        debugLog('Entity Notes: Observer not found, triggering fallback initialization...');
-        initializeWithErrorHandling();
-    } else {
-        debugLog('Entity Notes: Observer already exists, no fallback needed');
-    }
-}, 2000);
-
-infoLog('Entity Notes: Script loaded successfully');"""
-
     async def get(self, request):
         """Serve the JavaScript file."""
+        hass = request.app["hass"]
+        debug_logging = hass.data[DOMAIN]["config"].get(CONF_DEBUG_LOGGING, False)
+        max_note_length = hass.data[DOMAIN]["config"].get(CONF_MAX_NOTE_LENGTH, 200)
+        
+        # Get the JavaScript file path
+        js_file_path = Path(__file__).parent / FRONTEND_JS_PATH
+        
         try:
-            _LOGGER.debug("Serving embedded JavaScript content")
-            return web.Response(text=self.JS_CONTENT, content_type="application/javascript")
+            with open(js_file_path, 'r') as f:
+                js_content = f.read()
+                
+            # Replace configuration placeholders
+            js_content = js_content.replace('{{DEBUG_LOGGING}}', str(debug_logging).lower())
+            js_content = js_content.replace('{{MAX_NOTE_LENGTH}}', str(max_note_length))
+                
+            return web.Response(text=js_content, content_type='application/javascript')
             
+        except FileNotFoundError:
+            _LOGGER.error("JavaScript file not found: %s", js_file_path)
+            return web.Response(text="// Entity Notes: JavaScript file not found", content_type='application/javascript', status=404)
         except Exception as e:
             _LOGGER.error("Error serving JavaScript file: %s", e)
-            return web.Response(text="// Error loading JavaScript file", content_type="application/javascript", status=500)
-
+            return web.Response(text="// Entity Notes: Error loading script", content_type='application/javascript', status=500)
