+++ custom_components/entity_notes/const.py
@@ -1,13 +1,15 @@
 
 """Constants for Entity Notes component."""
-import logging
-
-_LOGGER = logging.getLogger(__name__)
 
 DOMAIN = "entity_notes"
 STORAGE_VERSION = 1
-STORAGE_KEY = "entity_notes"
+STORAGE_KEY = "entity_notes.notes"
 MAX_NOTE_LENGTH = 200
+
+# Configuration keys
+CONF_DEBUG_LOGGING = "debug_logging"
+CONF_MAX_NOTE_LENGTH = "max_note_length"
+CONF_AUTO_BACKUP = "auto_backup"
 
 # Events
 EVENT_NOTES_UPDATED = "entity_notes_updated"
@@ -17,6 +19,13 @@ SERVICE_SET_NOTE = "set_note"
 SERVICE_GET_NOTE = "get_note" 
 SERVICE_DELETE_NOTE = "delete_note"
 SERVICE_LIST_NOTES = "list_notes"
+SERVICE_BACKUP_NOTES = "backup_notes"
+SERVICE_RESTORE_NOTES = "restore_notes"
 
-# Module-level debug log to see if this file is being loaded
-_LOGGER.warning("🔧 DEBUG: const.py module loaded with DOMAIN: %s", DOMAIN)
+# Default configuration values
+DEFAULT_DEBUG_LOGGING = False
+DEFAULT_MAX_NOTE_LENGTH = 200
+DEFAULT_AUTO_BACKUP = True
+
+# File paths
+FRONTEND_JS_PATH = "entity-notes.js"
